Updated SDRAM to work with 32MB RAM (implemented A12 line); attempts to get SPI DMA working for LED controller

+++ Retrospector_H743_v2/src/Bootloader.cpp
@@ -2,7 +2,7 @@
 
 extern USB usb;
 extern SerialHandler serial;
-extern uint16_t adcZeroOffset[2];
+extern int32_t adcZeroOffset[2];
 
 #define BL_SAMPLE_SIZE 20000
 //uint8_t __attribute__((section (".sdramSection"))) bootloaderSamples[BL_SAMPLE_SIZE];
